move serialization fn to requirements module
