+++ src/python.rs
@@ -2,9 +2,7 @@ use crate::{check_output, check_status, is_not_found, relative_path, EResult, Pa
 use crate::{PathBuf, ProjectDirs};
 use eyre::eyre;
 use eyre::Context;
-use std::fmt::Write;
 
-use self::requirements::Requirement;
 use self::wheel::RecordEntry;
 
 pub(crate) mod detection;
@@ -136,24 +134,6 @@ pub(crate) fn records(
 
             (!is_dist_info).then(|| record.deserialize(None))
         }))
-}
-
-pub(crate) fn serialize_requirements_txt(reqs: &[Requirement]) -> String {
-    let mut output = String::new();
-    for req in reqs {
-        let _ = write!(&mut output, "{}=={}", req.name, req.version);
-        if let Some(marker) = req.marker.as_ref() {
-            let _ = write!(&mut output, "; {}", marker);
-        }
-        let _ = writeln!(&mut output, " \\");
-        let hashes = req
-            .available_hashes
-            .iter()
-            .map(|hash| format!("    --hash={}", hash.0.replace("=", ":")))
-            .collect::<Vec<_>>();
-        let _ = writeln!(&mut output, "{}", hashes.join(" \\\n"));
-    }
-    output
 }
 
 #[derive(PartialEq, Eq, Debug)]
